+++ core/shared/platform/windows/win_socket.c
@@ -86,7 +86,7 @@ fail:
 }
 
 int
-os_socket_settimeout(bh_socket_t socket, unsigned long long int timeout_us)
+os_socket_settimeout(bh_socket_t socket, uint64 timeout_us)
 {
     DWORD tv = (DWORD)(timeout_us / 1000UL);
